# Conflicts:
#	back_end_soruce/src/main/java/com/ateam/calmate/config/WebConfig.java
#	back_end_soruce/src/main/resources/application.yml

+++ back_end_soruce/src/main/java/com/ateam/calmate/report/command/dto/ReportCreateRequest.java
@@ -8,25 +8,11 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ReportCreateRequest {
-
-    // 신고 제목
     private String title;
-
-    // 신고 내용
     private String contents;
-
-    // 신고 당한 회원 번호 (report.member_id2)
-    private Long reportedMemberId;
-
-    // 신고한 회원 번호 (report.member_id)
-    private Long reporterMemberId;
-
-    // 신고 출처가 게시글이면 게시글 번호
-    private Integer postId;        // nullable
-
-    // 신고 출처가 댓글이면 댓글 번호
-    private Integer commentId;     // nullable
-
-    // 신고 사유 코드 (report_base.id → report.report_id)
-    private Integer reportBaseId;
+    private Long reportedMemberId;   // 피신고자
+    private Long reporterMemberId;   // 신고자
+    private Integer postId;          // nullable
+    private Integer commentId;       // nullable
+    private Long reportBaseId;       // report_base.id
 }
